+++ src/components/Text/index.tsx
@@ -1,7 +1,10 @@
+import { classNames } from '@utils/classNames';
+
 import styles from './index.module.css';
 
 interface IText {
     children: string;
+    className?: string;
 }
 
-export const Text = ({ children }: IText) => <p className={styles.container}>{children}</p>;
+export const Text = ({ children, className }: IText) => <p className={classNames(styles.container, className)}>{children}</p>;
