--- src/pages/Dashboard/parts/List/parts/Item/styles.module.css
@@ -1,13 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-}
-
-.link {
-    text-decoration: none;
-}
-
-.actions-wrapper {
-    display: flex;
-    justify-content: flex-end;
-}
